requirements/base.txt
pyaml
gunicorn
django
dpath
python-consul
redis
